finalize the fallback
